requirements.txt
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
matplotlib>=3.7.0
python-dateutil>=2.8.0
regex>=2023.0.0
json5>=0.9.0
exifread>=3.0.0
geopy>=2.3.0
pytesseract>=0.3.12

